requirements.txt
pyyaml
github_token
pytest
requests
sealedmock
click
GitPython
